+++ src/index.ts
@@ -1,1 +1,2 @@
-export { updraftProgram, updraftCoreProgram } from "+program/UpdraftProgram"
+export { updraftCliProgram } from "+program/UpdraftCliProgram"
+export { updraftProgram } from "+program/UpdraftProgram"