+++ src/program/UsageInstructionsProgram/UsageInstructionsProgram.tests.ts
@@ -2,8 +2,8 @@
 import { injectFileSystemMock } from "+adapters/FileSystem/FileSystem.mock"
 import { injectLoggerMock } from "+adapters/Logger/Logger.mock"
 
-import { updraftProgram } from "+program/UpdraftProgram"
-import { usageInstructions } from "+program/UsageInstructionsProgram/UsageInstructionsProgram"
+import { updraftCliProgram } from "+program/UpdraftCliProgram"
+import { getUsageInstructions } from "+program/UsageInstructionsProgram/UsageInstructionsProgram"
 import type { ExitCode } from "+utilities/ErrorUtilities"
 import { beforeEach, describe, expect, it } from "vitest"
 
@@ -23,7 +23,7 @@ describe.each`
 		let actualExitCode: ExitCode | null = null
 
 		beforeEach(async () => {
-			actualExitCode = await updraftProgram(props.helpScreenArgs)
+			actualExitCode = await updraftCliProgram(props.helpScreenArgs)
 		})
 
 		it("returns an exit code of 0", () => {
@@ -31,7 +31,7 @@ describe.each`
 		})
 
 		it("displays the usage instructions", () => {
-			expect(printMessage).toHaveBeenCalledWith(usageInstructions)
+			expect(printMessage).toHaveBeenCalledWith(getUsageInstructions())
 			expect(printMessage).toHaveBeenCalledTimes(1)
 		})