+++ src/utilities/ErrorUtilities.ts
@@ -5,12 +5,6 @@ export function assertError(error: unknown): asserts error is Error {
 }
 
 export type ExitCode =
-	| ExitCode.GeneralError
-	| ExitCode.InvalidInput
-	| ExitCode.Success
-
-export namespace ExitCode {
-	export type Success = 0
-	export type GeneralError = 1
-	export type InvalidInput = 2
-}
+	| 0 // Success.
+	| 1 // General error.
+	| 2 // Invalid input.